side bar remains in 404
