+++ fe/app/routes/portfolio.statistics_.tsx
@@ -1,0 +1,8 @@
+export default function Statistics() {
+    return (
+        <div>
+            <h1>Statistics</h1>
+            <p>View your statistics here</p>
+        </div>
+    );
+}
