+++ fe/app/routes/portfolio.transactions_.tsx
@@ -1,0 +1,8 @@
+export default function Transactions() {
+    return (
+        <div>
+            <h1>Transactions</h1>
+            <p>View your transactions here</p>
+        </div>
+    );
+}
