+++ fe/app/routes/portfolio_.tsx
@@ -1,0 +1,8 @@
+export default function Portfolio() {
+    return (
+        <div>
+            <h1>Portfolio</h1>
+            <p>Welcome to your stock portfolio</p>
+        </div>
+    );
+}
